Migrate Product/Order monetary fields from Decimal to Int (Task 12C.1)
Rial has no fractional subunit in practical use, so Product.price and the
Order/OrderItem money fields move from Decimal(10,2) to Int. The migration
casts existing values via ROUND (not truncate) and logs any row that had a
nonzero fractional part before rounding for manual review. All Decimal
arithmetic and toNumber()/Number() conversions on these fields are replaced
with plain number math.

+++ lib/validation/product.schema.ts
@@ -16,7 +16,7 @@ export const productCreateSchema = z.object({
     .regex(/^[A-Za-z0-9_-]+$/, "SKU may only contain letters, numbers, - and _"),
   categoryId: z.string().min(1),
   brandId: z.string().min(1),
-  price: z.number().min(0),
+  price: z.number().int().min(0),
   discountPercent: z.number().int().min(0).max(100).default(0),
   tags: tagsSchema,
   shortDescription: freeTextSchema(1, 300),
